+++ cider/serialization/src/lib.rs
@@ -1,0 +1,2 @@
+mod data_dump;
+pub use data_dump::*;
